fixed server group rename caused errors fixed cookie handling was broken in some cases changed message type now properly handles extra types add preliminary parsing for "plugin" messages
git-svn-id: http://miranda.googlecode.com/svn/trunk@1151 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/protocols/IcqOscarJ/icq_directmsg.c
@@ -381,7 +381,7 @@ void handleDirectGreetingMessage(directconnect* dc, PBYTE buf, WORD wLen, WORD w
 		{
 			icq_sendAdvancedMsgAck(dc->dwRemoteUin, 0, 0, wCookie, (BYTE)typeId, 0);
 		}
-		handleMessageTypes(dc->dwRemoteUin, time(NULL), 0, 0, wCookie, (BYTE)typeId, 0, 0, dwLengthToEnd, (WORD)dwDataLength, buf);
+		handleMessageTypes(dc->dwRemoteUin, time(NULL), 0, 0, wCookie, typeId, 0, 0, dwLengthToEnd, (WORD)dwDataLength, buf);
 	}
 	else
 	{
